pyproject.toml
[project]
name = "eventron"
version = "0.1.0"
description = "Eventron — 会场智能排座 Multi-Agent 系统"
requires-python = ">=3.10"
dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",

    # Agent orchestration
    "langgraph>=0.4",
    "langchain-core>=0.3",
    "langchain-openai>=0.3",
    "langchain-anthropic>=0.3",
    # "langchain-deepseek",  # uncomment when available on PyPI

    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "greenlet>=3.0",

    # Cache
    "redis>=5.0",

    # LangGraph checkpoint
    "langgraph-checkpoint-postgres>=2.0",

    # Excel
    "openpyxl>=3.1",

    # PDF / Badge
    "weasyprint>=62",
    "jinja2>=3.1",

    # QR Code
    "qrcode[pil]>=8",

    # Config
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",

    # HTTP client
    "httpx>=0.27",

    # JWT auth
    "pyjwt>=2.8",

    # Email validation
    "email-validator>=2.0",

    # Utilities
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "testcontainers[postgres]>=4.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "agents", "tools"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short -p no:cacheprovider"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.isort]
known-first-party = ["app", "agents", "tools"]

[tool.mypy]
python_version = "3.10"
plugins = ["pydantic.mypy"]
warn_return_any = true
warn_unused_configs = true
